--- controllers/contacts/contacts-controllers.js
@@ -1,86 +1,0 @@
-const { ctrlWrapper } = require("../../utils");
-
-const { Contact } = require("../../models/contact");
-
-const { HttpError } = require("../../helpers");
-
-const getAllContacts = async (req, res) => {
-  const { _id: owner } = req.user;
-  //Реалізовуємо пагінацію
-  const { page = 1, limit = 10 } = req.query;
-  const skip = (page - 1) * limit;
-  // const { favorite } = req.query;
-
-  const result = await Contact.find({ owner }, "-createdAt -updatedAt", {
-    skip,
-    limit,
-  }).populate("owner", "email subscription");
-  res.json(result);
-};
-
-const getContactById = async (req, res) => {
-  const { contactId } = req.params;
-
-  const result = await Contact.findById(contactId);
-
-  if (!result) {
-    throw HttpError(404);
-  }
-  res.json(result);
-};
-
-const addContact = async (req, res) => {
-  //Забираємо через деструктиризацію з об'єкту запиту (req) "_id" та відразу переіменовуємо його в "owner"
-  const { _id: owner } = req.user;
-
-  const result = await Contact.create({ ...req.body, owner });
-  res.status(201).json(result);
-};
-
-const deleteContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndDelete(contactId);
-
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-
-  res.status(200).json({
-    message: "contact deleted",
-  });
-};
-
-const updateContactById = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-
-  if (!result) {
-    throw HttpError(404, `Contact with ${contactId} not found`);
-  }
-
-  res.json(result);
-};
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-
-  if (!result) {
-    throw HttpError(404, `"message": "Not found"`);
-  }
-
-  res.json(result);
-};
-
-module.exports = {
-  getAllContacts: ctrlWrapper(getAllContacts),
-  getContactById: ctrlWrapper(getContactById),
-  addContact: ctrlWrapper(addContact),
-  deleteContact: ctrlWrapper(deleteContact),
-  updateContactById: ctrlWrapper(updateContactById),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-};
