+++ controllers/contacts/deleteContact.js
@@ -1,0 +1,18 @@
+const { Contact } = require("../../models/contact");
+
+const { HttpError } = require("../../helpers");
+
+const deleteContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndDelete(contactId);
+
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+
+  res.status(200).json({
+    message: "contact deleted",
+  });
+};
+
+module.exports = deleteContact;
